+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.25.0
+
+- **Add** : Added Klarna payment method
+- **Fix** : Fixed issues on order timeout by redirecting customer to pending URL
+- **Fix** : Fixed stock discount when *Blocked* (110) or *Denied* (111) notification is received
+
 ## 1.24.2
 
 - **Fix** : Fixed display issue about ApplePay payment method which was displayed on unsupported devices